move entry helpers to lib file
